# backend/openapi.yaml
openapi: 3.0.3
info:
  title: Cassandra API
  version: 1.0.0
  description: REST API per il monitoraggio smart city di Trento. Fornisce dati in tempo reale su meteo, traffico, illuminazione e folla, oltre alla gestione utenti e pannello admin.
servers:
  - url: http://localhost:3000
    description: Sviluppo locale
  - url: https://cassandra-backend.onrender.com
    description: Produzione (Render.com)
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: 'JWT ottenuto dal login. Header: Authorization: Bearer <token>'
  schemas:
    AuthResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
        token:
          type: string
          description: JWT valido per 24 ore
        id:
          type: string
        name:
          type: string
          example: Mario Rossi
        email:
          type: string
          format: email
        role:
          type: string
          enum:
            - user
            - admin
        self:
          type: string
          example: /api/v1/users/64abc...
...
    Error:
      type: object
      properties:
        success:
          type: boolean
          example: false
        message:
          type: string
paths:
  /api/v1/authentications:
    post:
      tags:
        - Auth
      summary: Login con email e password
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - email
                - password
              properties:
                email:
                  type: string
                  format: email
                  example: mario@example.com
                password:
                  type: string
                  example: user123
      responses:
        '200':
          description: Login effettuato
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        '400':
          description: Campi mancanti o account Google-only
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Credenziali non valide
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
  /api/v1/routing:
    get:
      tags:
        - Routing
      summary: Calcola il percorso stradale tra due punti
      description: Restituisce distanza, durata e geometria del tragitto in auto tra origine e destinazione, calcolati tramite OSRM. Tra i percorsi alternativi viene scelto quello suggerito meno volte, per non sovraffollare sempre le stesse strade.
      parameters:
        - name: fromLat
          in: query
          required: true
          schema:
            type: number
          example: 46.0723
          description: Latitudine di partenza
        - name: fromLon
          in: query
          required: true
          schema:
            type: number
          example: 11.1162
          description: Longitudine di partenza
        - name: toLat
          in: query
          required: true
          schema:
            type: number
          example: 46.066
          description: Latitudine di destinazione
        - name: toLon
          in: query
          required: true
          schema:
            type: number
          example: 11.1498
          description: Longitudine di destinazione
      responses:
        '200':
          description: Percorso calcolato
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  distance:
                    type: integer
                    description: Distanza in metri
                    example: 3200
                  duration:
                    type: integer
                    description: Durata in secondi
                    example: 540
                  geometry:
                    type: object
                    description: GeoJSON LineString, coordinate [lon, lat]
                    properties:
                      type:
                        type: string
                        example: LineString
                      coordinates:
                        type: array
                        items:
                          type: array
                          items:
                            type: number
                  alternatives:
                    type: integer
                    description: Numero di percorsi tra cui è stato scelto
                    example: 2
        '400':
          description: Coordinate mancanti o non valide
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Nessun percorso trovato tra i due punti
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '503':
          description: Servizio di routing non disponibile
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
  /api/v1/admin/users:
    get:
      tags:
        - Admin
      summary: Lista tutti gli utenti registrati
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Lista utenti senza password